// frontend/src/app/model/contact.model.ts
export class Contact {
  firstname: string;
  lastname: string;
  phone: string;
  address1: string;
  address2: string;
  state: string;
  city: string;
  zip: string;

  constructor(firstname: string,
    lastname: string,
    phone: string,
    address1: string,
    address2: string,
    state: string,
    city: string,
    zip: string) {}

}